[Android] Fixes Up the Gradle Version And RN Maven Path
Summary:
I believe starting with RN 21 (maybe), the dependencies are no longer
on maven / mavencentral but instead you reference them from the
local node_modules folder.
Second in order to make the Android lib project work with AS, we have
to bump the gradle version. It now matches the version in the RN hello world project.

: